requirements.txt
Flask==2.1.2
Flask-MySQL==1.5.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
Flask-Login==0.6.2
Flask-Cors==2.1.0
requests==2.26.0
gunicorn==20.1.0
Flask-Session==0.4.0
flask-mail==0.9.1
pandas==1.4.2
json5==0.9.9
Werkzeug==2.1.2
boto3==1.28.10
python-dotenv==0.21.0
awscli==1.29.25
mysql-connector-python
google-generativeai
pymysql
